Feat: Store the food chosen by the user in local storage
